Always include a proxy header in the request body.
Previously only AMQP and COAP included the proxy header. Now also
include the proxy header in HTTP requests, although it is not
actually needed since the properties can be passed in the HTTP headers.

The reason for this change is to not to distinguish between different
protocols. This makes it much easier to do the same processing on
exchanges originating from different protocols.

In specefic this is due to the way we deal with GZIPing.
Since we now enforce GZIPing on ALL exchanges, passing an empty body
will now cause an exception. This is caused by the third party code,
faulting if no body is present.